+++ modules/clawbot_server/main.tf
@@ -38,6 +38,8 @@ locals {
       enable_root_ssh                       = var.enable_root_ssh
       openclaw_repo_url                     = var.openclaw_repo_url
       openclaw_branch                       = var.openclaw_branch
+      openclaw_agent_pack_repo_url          = var.openclaw_agent_pack_repo_url
+      openclaw_agent_pack_ref               = var.openclaw_agent_pack_ref
       openclaw_gateway_token                = var.openclaw_gateway_token
       openclaw_opt_volume_enabled           = var.opt_volume_enabled
       openclaw_opt_volume_fstype            = var.opt_volume_fstype